Move weight/bias preparation implementation to new file
